nii2pptx/requirements.txt
python-pptx
matplotlib
nibabel
numpy
Pillow
imageio
tk
PyMuPDF  
Pillow 
opencv-python
pywin32